themes
- renaming folder theme to themes (dark and light)
- moved universal component styles to another file
